+++ tests/testthat/test_Resampling.R
@@ -41,3 +41,24 @@ test_that("hashing", {
     expect_false(identical(r$hash, hash))
   }
 })
+
+test_that("integer grouping col (#396)", {
+  df = data.frame(
+    id = rep(1L:10L, each = 2),
+    x = rnorm(20)
+  )
+
+  tsk = TaskRegr$new(id = "task", backend = df, target = "x")
+  tsk$set_col_role("id", "group")
+
+  bs = rsmp("bootstrap", repeats = 10L, ratio = 1)
+  bs$instantiate(tsk)
+
+  set = bs$train_set(1)
+  expect_integer(set)
+  expect_true(all(map_lgl(split(seq_row(df), f = df$id), function(x) all(x %in% set) || all(x %nin% set))))
+
+  set = bs$test_set(1)
+  expect_integer(set)
+  expect_true(all(map_lgl(split(seq_row(df), f = df$id), function(x) all(x %in% set) || all(x %nin% set))))
+})